Move tests next to implementations
