+++ src/polynomial-features.test.ts
@@ -1,6 +1,6 @@
-import { PolynomialFeatures } from '../index.js'
-import { RegressionError } from '../util/util.js';
-import { PolynomialFeaturesConfig } from '../polynomial-features.js';
+import { PolynomialFeatures } from './index.js'
+import { RegressionError } from './util/util.js';
+import { PolynomialFeaturesConfig } from './polynomial-features.js';
 
 /// Auxiliary function
