+++ tsconfig.build.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.json",
   "exclude": [
-    "src/__tests__"
+    "src/**/*.test.ts"
   ]
 }
