+++ tsconfig.json
@@ -18,6 +18,6 @@
     "exactOptionalPropertyTypes": true,
   },
   "include": [
-    "src"
+    "src/**/*"
   ]
 }
